requirements.txt
bcrypt==4.2.1
numpy==2.2.3
pandas==2.2.3
pip==25.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
wheel==0.45.1
